editor tests
